# docs/DigiShield_openapi.yaml
openapi: 3.0.3
info:
  title: DigiShield (Digital Shield) API
  version: 1.0.0
  description: >
    Platform API for prevention, training & alerting against online scams in cyberspace.
    Multi-tenant, RBAC, scoped by org_id in the JWT. Errors follow RFC 7807.
  contact:
    name: DigiShield Engineering
servers:
  - url: https://api.digishield.vn/api/v1
    description: Production
  - url: https://staging-api.digishield.vn/api/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List groups (including smart groups)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
    post:
      tags: [Groups]
      summary: Create group (rule_json for smart group)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Group' }
      responses:
        '201': { description: Created }
  /groups/{id}/evaluate:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Groups]
      summary: Re-evaluate smart group members
      responses:
        '200':
          description: Member count after evaluation
          content:
            application/json:
              schema:
                type: object
                properties:
                  member_count: { type: integer }
  /groups/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    patch:
      tags: [Groups]
      summary: Update a group
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Group' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
    delete:
      tags: [Groups]
      summary: Delete a group
      responses:
        '204': { description: Deleted }
  /groups/{id}/members/{userId}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Groups]
      summary: Remove a member from a group
      responses:
        '200':
          description: Member count after the change
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberCount' }
  /groups/{id}/members:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Groups]
      summary: Member ids of a group
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string, format: uuid }
    post:
      tags: [Groups]
      summary: Add a member to a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        '200':
          description: Member count after the change
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberCount' }
...
  /tenants:
    get:
      tags: [Tenancy]
      summary: List tenants (Super Admin)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tenant' }
    post:
      tags: [Tenancy]
      summary: Create a new tenant, and its first administrator
      description: >-
        Users are always created in the caller's own tenant, so adminEmail is the
        only moment an account can be placed inside the new one. Omit it and the
        organisation exists with nobody able to sign in. Provisioning the admin
        happens in the same transaction: if it fails, the tenant is not created.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                tier:
                  type: string
                  enum: [pool, bridge, silo]
                dataRegion:
                  type: string
                adminEmail:
                  type: string
                  format: email
                  description: first administrator; receives the invitation email
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
  /tenants/{id}/impersonate:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Tenancy]
      summary: Enter a tenant as a platform operator
      description: >-
        Writes the entry into that tenant's own audit log, so impersonation
        leaves a trace the customer can see (business rule BR-6).
      responses:
        '204': { description: Recorded }
  /tenants/{id}/thresholds:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Tenancy]
      summary: Business thresholds of a tenant
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessThresholds' }
    patch:
      tags: [Tenancy]
      summary: Update business thresholds
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessThresholds' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessThresholds' }
  /tenants/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Tenancy]
      summary: Tenant details
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
    patch:
      tags: [Tenancy]
      summary: Update tenant (tier, status, data region)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
  /tenants/{id}/settings:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Tenancy]
      summary: Tenant configuration (branding, policy, language)
      description: >-
        A tenant with no identity provider connected still answers 200, with
        connected=false and syncStatus=never. 404 means the tenant itself does
        not exist.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScimConfig' }
        '404':
          description: No such tenant
    patch:
      tags: [Tenancy]
      summary: Update tenant configuration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantSettings' }
      responses:
        '200': { description: OK }
...
  /tenants/{id}/feature-flags:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Tenancy]
      summary: List tenant feature flags
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FeatureFlag' }
    patch:
      tags: [Tenancy]
      summary: Enable/disable feature flag
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeatureFlag' }
      responses:
        '200': { description: OK }
  /tenants/{id}/usage:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Billing]
      summary: Tenant usage metering data
      parameters:
        - in: query
          name: period
          schema: { type: string, example: 2026-06 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UsageMetering' }
  /tenants/{id}/subscription:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Billing]
      summary: Tenant's current subscription
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
    put:
      tags: [Billing]
      summary: Change subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
  /audit:
    get:
      tags: [Tenancy]
      summary: Audit log of the current tenant
      description: Also served at /super/audit for platform operators.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditLogEntry' }
        '403':
          description: Requires ORG_ADMIN or above
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Group:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        rule_json:
          type: object
          description: Dynamic conditions for the smart group
          example: { risk_score_gte: 70, department: "ke-toan" }
        member_count: { type: integer }
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        name: { type: string }
        tier: { type: string, enum: [pool, bridge, silo] }
        dataRegion: { type: string, example: vn }
        status: { type: string, enum: [provisioning, active, suspended, deactivated] }
        userCount: { type: integer }
        domain: { type: string }
    TenantUpdate:
      type: object
      properties:
        name: { type: string }
        tier: { type: string, enum: [pool, bridge, silo] }
        status: { type: string, enum: [provisioning, active, suspended, deactivated] }
        data_region: { type: string, example: vn }
    TenantSettings:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        branding: { type: object, description: logo, colors, subdomain }
        policy: { type: object, description: risk thresholds, mandatory training }
        default_locale: { type: string, example: vi }
    FeatureFlag:
      type: object
      properties:
        key: { type: string, example: deepfake_sim }
        enabled: { type: boolean }
...
    Subscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        plan_id: { type: string, format: uuid }
        status: { type: string, enum: [trial, active, past_due, canceled] }
        renews_at: { type: string, format: date }
...
    AuditLogEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        ts: { type: string, format: date-time }
        actor: { type: string }
        action: { type: string }
        target: { type: string }
        ip: { type: string }
        severity: { type: string }
...
    UsageMetering:
      type: object
      properties:
        tenant_id: { type: string, format: uuid }
        metric: { type: string, enum: [email_sent, sms_sent, ai_call, storage] }
        value: { type: integer }
        period: { type: string, example: 2026-06 }
...
    MemberCount:
      type: object
      properties:
        member_count: { type: integer }
    BusinessThresholds:
      type: object
      properties:
        risk_alert_score: { type: integer, minimum: 0, maximum: 100 }
        pass_score_pct: { type: integer, minimum: 0, maximum: 100 }
        min_campaigns_per_quarter: { type: integer, minimum: 0 }
    ScimConfig:
      type: object
      properties:
        tenantId: { type: string, format: uuid }
        idpName: { type: string }
        connected: { type: boolean }
        idpTenantId: { type: string }
        clientId: { type: string }
        scimEndpoint: { type: string }
        lastSyncAt: { type: string, format: date-time }
        syncedUserCount: { type: integer }
        syncErrorCount: { type: integer }
        syncStatus: { type: string }
...